Share writer's array of in-progress subtransactions with readers.
Including subtransaction array (subxip) in the shared snapshot allows readers
to use it in much the same way as in-progress transactions array (xip).
Readers can now use it to determine if a committed subtransaction's effects are
visible or not.  Previously, in-progress subtransactions were not included in
the snapshot shared with readers and that broke visibility.

Fixes GitHub issue #3528.