query_finish_pending: reset a fault early to avoid interference
The query_finish_pending test leaves execsort_mksort_mergeruns fault
injected even after it is triggered once.  This seems unintentional and
it interferes with the subsequent execshare_input_next, because both the
faults happen to be covered by the SQL statement under test.  The fault
status of execshare_input_next may be returned as not-triggered, causing
the test to fail.  Fix this by removing the mergeruns fault as soon as
it's triggered once.