New gp_toolkit functions to get and remove entries from AppendOnlyHash
The functions allow obtaining or removing entries from the shared hash
table maintained on QD.  Default size of this hash table is 1000 and
entries are removed only after it is filled to capacity.  The two
functions should be helpful for testing as well as troubleshooting
issues with appendonly tables in production deployments.

Co-authored-by: Jimmy Yih <jyih@pivotal.io>