WIP: Fix use-after-free bug in memtuples.
The bug manifests only when a datum is detoasted and the slot
containing this tuple is referenced again.  Unique node in executor is
one example where this happens - the previous slot needs to be
compared with the current slot for uniqueness.

Next: find a failing test in master and fix in master first.

Co-authored-by: David Kimura <dkimura@pivotal.io>