Test using psql variables to promote standby
This is just to demonstrate how tricky it is to get standby promoted
and then to restore the state back to where we came from.  Isolation2
framework doesn't support psql variables.  We can replace them with,
say, plpgsql functions.  However, isolation2 framework does not deal
very well with disconnections from current master and reconnections to
it later in the same spec...