Protect XIP array in reader's snapshot
QE reader obtains its snapshot by copying it from its writer process.  In order
to debug bogus values in reader's snapshot XIP array, this patch marks the
memory region as read-only as soon as it's copied from the writer's snapshot.
If there is a logical error where the XIP array is accidentally written to
afterwards, this commit should catch it.