Test that replay lag does not affect streaming replication
The test slows down the replay process using a sleep fault.  Then it
breaks the streaming WAL replication connection using another fault.
The connection should be restored automatically.  The test makes a
commit and checks if it waits for syncrep.  If the commit waits for
syncrep, the test fails.  Currently, the design is such that, WAL
receiver process waits for WAL replay process for the streaming start
location.  That causes the test to fail.  Ideally, WAL receiver process
should be able to identify the streaming start location without
depending on the replay process.  If that happens, the commits will not
get stuck until the end of pending WAL replay and the test would pass.