Start WAL receiver as early as possible
This patch avoids unexpected behavior when WAL replication connection
breaks (e.g. due to network blip or standby shutdown) and WAL receiver
needs to be restarted.  Specifically, if replay lag is built up, WAL
receiver used to restart only after startup process finished replaying
all WAL already received.  This is now changed such that before
replaying each WAL record, the startup process checks if WAL receiver
needs to be started.

A new GUC is introduced to control this behavior.  By default the new
change does not take effect.

Each time a new WAL segment is created by WAL receiver, it remembers
the start LSN on the new segment in recovery.conf.  This is used to
during startup to determine the where to start streaming from.

This patch is based on the pgsql-hackers discussions (and patch):
https://www.postgresql.org/message-id/CANXE4TewY1WNgu5J5ek38RD%2B2m9F2K%3DfgbWubjv9yG0BeyFxRQ%40mail.gmail.com
https://www.postgresql.org/message-id/b271715f-f945-35b0-d1f5-c9de3e56f65e@postgrespro.ru