WIP: Raise error when affecting tuple moved into another segment
Split update (updating a distribution key column) in Greenplum is
implemented by delete followed by insert.  This is similar to an
update to partition key column in upstream PostgreSQL, which is
handled by commit f16241bef7cc.  This patch cherry-picks relevant
parts of that commit.  The idea is to distinguish a tuple deleted as
part of split update from a regular deleted tuple.  Concurrent
attempts to update or delete such a tuple need to be aware that the
update chain for this tuple cannot be traced because the updated tuple
no longer resides in the same heap table (it's moved to another
segment).

Main difference between the upstream commit and this patch is the
tuple header change is not recorded in WAL.  Mirrors see this as
regular heap delete during replay and crash recovery would also replay
it as regular heap delete.  That is, the tuple header change is
treated like a hint bit change.  Tuple masking logic used by replica
check and WAL consistency checks is updated to mask the tuple header
change.  We think this is OK in Greenplum 6 because hot standby does
not work here.

TODO:
1. port the comment in htup_details.h
2. port the isolation tests from commit f16241bef7cc, replace
   'partition by' with 'distributed by'
3. add tests from PR #9461

Co-authored-by: Zhenghua Lyu <kainwen@gmail.com>