Allow CLUSTER on append-optimized tables
Cluster on AO tables is implemented by sorting the entire AO table using
tuple sort framework, according to a btree index defined on the table.

A faster way to cluster is to scan the tuples in index-order, but this
requires index-scan support.  Append-optimized tables do not support
index-scans currently, but when this support is added, the cluster
operation can be enhanced accordingly.

Author: Andrey Borodin <amborodin@acm.org>
Reviewed and slightly edited by: Asim R P <pasim@vmare.com>

Merges GitHub PR #9996