Implement storage manager interface for AO tables
This started with an attempt to fix fsync_ao test failure.  The patch
includes a new implementation of storage manager API that can be used
for relations that have the following characteristics:

1. shared buffer cache is not used, kernel page cache is relied upon

2. unlike heap, segment files need not be identified with consecutive
   integers

3. variable length blocks are writtent to segment files

There are several problems with the implementation in this patch,
however, it is a step in the right direction.  FWIW, the fsync_ao test
passes with this patch.

Co-authored-by: Georgios Kokolatos <gkokolatos@pivotal.io>