Determine gp_role dynamically, simplify startup sequence
We should be able to use the native "pg_ctl start" command to start a
Greenplum segment.  The postmaster should be able to determine whether
it is a coordinator, standby, primary, or mirror.  Accordingly, it
should be able to set gp_role properly.

The patch attempts to achieve this goal.  It has a couple of benefits.
Greenplum specific command line options are no longer mandatory.
Also, pg_ctl no longer needs to be aware of Greenplum specific
"dtmready" state.  This allows third-party extensions that are written
for PostgreSQL to work seamlessly with Greenplum - case in point is
pg_auto_failover.

Second benefit is simplification of cluster startup sequence.  We no
longer need to start coordinator in utility mode, get the segment
details, shut it down and restart in normal mode after all segments
have started.  The coordinator's postmaster assumes gp_role as utility
upon startup.  A new UDF, "transition_to_dispatch_mode()" is
introduced as an interface for gpstart.  This UDF is invoked after
gpstart starts all the segments.

Next steps:

(1) Debug why gp_role remains "utility" after gpstart returns, it
    should be "dispatch".

(2) See if FTS can be made intelligent, so that it doesn't mark all
    primaries as down when during startup when they are not yet
    started.  With this, the "transition_to_dispatch_mode()" UDF can
    also be eliminated.

Co-authored-by: Hubert Zhang <zhubert@vmware.com>